save CubeQuery, get data, calcualte results

+++ database/migrations/2016_12_26_213631_create_cube_queries_table.php
@@ -22,7 +22,7 @@ class CreateCubeQueriesTable extends Migration
             $table->tinyInteger('x2')->nullable();
             $table->tinyInteger('y2')->nullable();
             $table->tinyInteger('z2')->nullable();
-            $table->tinyInteger('w')->nullable();
+            $table->float('w')->nullable();
             $table->float('result')->nullable();
             $table->unsignedInteger('cube_id');
             $table->foreign('cube_id')->references('id')->on('cubes');
